requirements.txt
streamlit>=1.32.0
opencv-python-headless>=4.9.0.80
Pillow>=10.2.0
matplotlib>=3.8.3
seaborn>=0.13.2
numpy>=1.26.4
pandas>=2.2.1
